feat: update login/account links when auth state changes
Add a default script that dispatches events for auth state changes.
Incorporate that into existing auth scripts and provide a default script
that handles it on pages that otherwise don't need auth.

Something to look out for: on pages that do need auth changes, they
should import and run the `observeAuthState` function. It won't be
loaded by default on auth pages, since we don't want to double up on
loading firebase.